Removed the tests that validate the properties because the validation moved to the batch wrapper Moved calc_image_properties to the module image properties ImagePropertyDrift refractored
