+++ deepchecks/vision/checks/train_test_validation/image_property_drift.py
@@ -22,7 +22,7 @@ from deepchecks.utils.dict_funcs import get_dict_entry_by_value
 from deepchecks.utils.distribution.drift import calc_drift_and_plot
 from deepchecks.utils.strings import format_number
 from deepchecks.vision import Batch, Context, TrainTestCheck
-from deepchecks.vision.utils.image_properties import default_image_properties, get_column_type, validate_properties
+from deepchecks.vision.utils.image_properties import default_image_properties, get_column_type, calc_image_properties
 
 __all__ = ['ImagePropertyDrift']
 
@@ -86,11 +86,7 @@ class ImagePropertyDrift(TrainTestCheck):
             **kwargs
     ):
         super().__init__(**kwargs)
-        if image_properties is not None:
-            self.image_properties = validate_properties(image_properties)
-        else:
-            self.image_properties = default_image_properties
-
+        self.image_properties = image_properties
         self.margin_quantile_filter = margin_quantile_filter
         if max_num_categories is not None:
             warnings.warn(
@@ -109,12 +105,16 @@ class ImagePropertyDrift(TrainTestCheck):
         self._train_properties = None
         self._test_properties = None
         self._class_to_string = None
+        self.properties_list = None
 
     def initialize_run(self, context: Context):
         """Initialize self state, and validate the run context."""
         self._class_to_string = context.train.label_id_to_name
+
         self._train_properties = defaultdict(list)
         self._test_properties = defaultdict(list)
+        self.properties_list = context.get_data_by_kind(DatasetKind.TRAIN).image_properties \
+            if self.image_properties is None else self.image_properties
 
     def update(
         self,
@@ -124,9 +124,9 @@ class ImagePropertyDrift(TrainTestCheck):
     ):
         """Calculate image properties for train or test batch."""
         if dataset_kind == DatasetKind.TRAIN:
-            properties = self._train_properties
+            properties_results = self._train_properties
         elif dataset_kind == DatasetKind.TEST:
-            properties = self._test_properties
+            properties_results = self._test_properties
         else:
             raise RuntimeError(
                 f'Internal Error - Should not reach here! unknown dataset_kind: {dataset_kind}'
@@ -143,9 +143,13 @@ class ImagePropertyDrift(TrainTestCheck):
                 if any(cls in map(self._class_to_string, classes[idx]) for cls in self.classes_to_display)
             ]
 
-        for single_property in self.image_properties:
-            property_list = single_property['method'](images)
-            properties[single_property['name']].extend(property_list)
+        data_for_properties = calc_image_properties(images, self.properties_list)
+
+        for prop_name, property_values in data_for_properties.items():
+            properties_results[prop_name].extend(property_values)
+        # for single_property in self.image_properties:
+        #     property_list = single_property['method'](images)
+        #     properties[single_property['name']].extend(property_list)
 
     def compute(self, context: Context) -> CheckResult:
         """Calculate drift score between train and test datasets for the collected image properties.
@@ -182,7 +186,7 @@ class ImagePropertyDrift(TrainTestCheck):
         drifts = {}
         not_enough_samples = []
 
-        for single_property in self.image_properties:
+        for single_property in self.properties_list:
             property_name = single_property['name']
 
             try:
