+++ deepchecks/vision/utils/image_properties.py
@@ -11,6 +11,7 @@
 """Module containing the image formatter class for the vision module."""
 import warnings
 from typing import Any, Dict, List, Tuple
+from collections import defaultdict
 
 import numpy as np
 from skimage.color import rgb2gray
@@ -213,3 +214,27 @@ def get_column_type(output_type):
         'categorical': 'categorical'
     }
     return mapper[output_type]
+
+def calc_image_properties(images, properties_to_calc) -> Dict[str, list]:
+    """
+    Calculates the image properties for a batch of images.
+
+    Parameters
+    ----------
+    images : torch.Tensor
+        Batch of images to transform to image properties.
+
+    image_properties: List[Dict] , default: None
+        overrides self.image_proerties, if None uses self.image properties
+
+    Returns
+    ------
+    batch_properties: dict[str, List]
+        A dict of property name, property value per image
+    """
+
+    batch_properties = defaultdict(list)
+    for single_property in properties_to_calc:
+        property_list = single_property['method'](images)
+        batch_properties[single_property['name']] = property_list
+    return batch_properties
