+++ tests/vision/checks/train_test_validation/image_property_drift_test.py
@@ -66,24 +66,6 @@ def test_image_property_drift_check_limit_classes_illegal(coco_train_visiondata,
     )
 
 
-def test_image_property_drift_initialization_with_empty_list_of_image_properties():
-    assert_that(
-        calling(ImagePropertyDrift).with_args(image_properties=[]),
-        raises(DeepchecksValueError, 'Properties list can\'t be empty')
-    )
-
-
-def test_image_property_drift_initialization_with_list_of_invalid_image_properties():
-    assert_that(
-        calling(ImagePropertyDrift).with_args(image_properties=[{'hello': 'string'}]),
-        raises(
-            DeepchecksValueError,
-            r"List of properties contains next problems:\n"
-            rf"\+ Property #0: dictionary must include keys \('name', 'method', 'output_type'\)\. "
-            fr"Next keys are missed \['method', 'name', 'output_type'\]")
-    )
-
-
 def test_image_property_drift_condition(coco_train_visiondata, coco_test_visiondata, device):
     result = (
         ImagePropertyDrift()
